requirements.txt
pandas
numpy
scikit-learn
joblib
streamlitpip